+++ Areas/Admin/Models/AdminMenu.cs
@@ -8,7 +8,7 @@ namespace DoAn_WebAcc.Areas.Admin.Models
 
     {
         [Key]
-        public long AdminMenuId { get; set; }
+        public int AdminMenuId { get; set; }
         public string ItemName { get; set; }
         public int ParentLevel { get; set; }
         public int ItemOrder { get; set; }
